Multi-Platform Link Crawling + Data Extraction System/requirements.txt
python-docx==1.0.1
PyPDF2==3.0.1
playwright==1.41.0
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.0
pytest==7.4.3
radon==6.0.1
gitpython==3.1.40
coverage==7.3.2
pydantic==2.6.1
aiohttp==3.9.1
lxml==4.9.3

pip install PyPDF2
pip install python-docx
pip install beautifulsoup4
pip install aiohttp
pip install requests